simplify bucket_gocb/collection diff
