+++ base/collection.go
@@ -13,17 +13,12 @@ licenses/APL2.txt.
 package base
 
 import (
-	"bytes"
 	"context"
 	"errors"
 	"expvar"
 	"fmt"
 	"io"
-	"math/rand"
 	"net/http"
-	"net/url"
-	"strconv"
-	"strings"
 	"sync"
 	"time"
 
@@ -622,228 +617,4 @@ func (b *GocbV2Bucket) waitForAvailQueryOp() {
 
 func (b *GocbV2Bucket) releaseQueryOp() {
 	<-b.queryOps
-}
-
-func createBatchesEntries(batchSize uint, entries []*sgbucket.BulkSetEntry) [][]*sgbucket.BulkSetEntry {
-	// boundary checking
-	if len(entries) == 0 {
-		WarnfCtx(context.Background(), "createBatchesEnrties called with empty entries")
-		return [][]*sgbucket.BulkSetEntry{}
-	}
-	if batchSize == 0 {
-		WarnfCtx(context.Background(), "createBatchesEntries called with invalid batchSize")
-		result := [][]*sgbucket.BulkSetEntry{}
-		return append(result, entries)
-	}
-
-	batches := [][]*sgbucket.BulkSetEntry{}
-	batch := []*sgbucket.BulkSetEntry{}
-
-	for idxEntry, entry := range entries {
-
-		batch = append(batch, entry)
-
-		isBatchFull := uint(len(batch)) == batchSize
-		isLastEntry := idxEntry == (len(entries) - 1)
-
-		if isBatchFull || isLastEntry {
-			// this batch is full, add it to batches and start a new batch
-			batches = append(batches, batch)
-			batch = []*sgbucket.BulkSetEntry{}
-		}
-
-	}
-	return batches
-
-}
-
-func createBatchesKeys(batchSize uint, keys []string) [][]string {
-
-	// boundary checking
-	if len(keys) == 0 {
-		WarnfCtx(context.Background(), "createBatchesKeys called with empty keys")
-		return [][]string{}
-	}
-	if batchSize == 0 {
-		WarnfCtx(context.Background(), "createBatchesKeys called with invalid batchSize")
-		result := [][]string{}
-		return append(result, keys)
-	}
-
-	batches := [][]string{}
-	batch := []string{}
-
-	for idxKey, key := range keys {
-
-		batch = append(batch, key)
-
-		isBatchFull := uint(len(batch)) == batchSize
-		isLastKey := idxKey == (len(keys) - 1)
-
-		if isBatchFull || isLastKey {
-			// this batch is full, add it to batches and start a new batch
-			batches = append(batches, batch)
-			batch = []string{}
-		}
-
-	}
-	return batches
-
-}
-
-// If the error is a net/url.Error and the error message is:
-//
-//	net/http: request canceled while waiting for connection
-//
-// Then it means that the view request timed out, most likely due to the fact that it's a stale=false query and
-// it's rebuilding the index.  In that case, it's desirable to return a more informative error than the
-// underlying net/url.Error. See https://github.com/couchbase/sync_gateway/issues/2639
-func isGoCBQueryTimeoutError(err error) bool {
-
-	if err == nil {
-		return false
-	}
-
-	// If it's not a *url.Error, then it's not a viewtimeout error
-	netUrlError, ok := pkgerrors.Cause(err).(*url.Error)
-	if !ok {
-		return false
-	}
-
-	// If it's a *url.Error and contains the "request canceled" substring, then it's a viewtimeout error.
-	return strings.Contains(netUrlError.Error(), "request canceled")
-
-}
-
-// putDDocForTombstones uses the provided client and endpoints to create a design doc with index_xattr_on_deleted_docs=true
-func putDDocForTombstones(name string, payload []byte, capiEps []string, client *http.Client, username string, password string) error {
-
-	// From gocb.Bucket.getViewEp() - pick view endpoint at random
-	if len(capiEps) == 0 {
-		return errors.New("No available view nodes.")
-	}
-	viewEp := capiEps[rand.Intn(len(capiEps))]
-
-	// Based on implementation in gocb.BucketManager.UpsertDesignDocument
-	uri := fmt.Sprintf("/_design/%s", name)
-	body := bytes.NewReader(payload)
-
-	// Build the HTTP request
-	req, err := http.NewRequest("PUT", viewEp+uri, body)
-	if err != nil {
-		return err
-	}
-	req.Header.Add("Content-Type", "application/json")
-	req.SetBasicAuth(username, password)
-
-	resp, err := client.Do(req)
-	if err != nil {
-		return err
-	}
-
-	defer ensureBodyClosed(resp.Body)
-	if resp.StatusCode != 201 {
-		data, err := io.ReadAll(resp.Body)
-		if err != nil {
-			return err
-		}
-		return fmt.Errorf("Client error: %s", string(data))
-	}
-
-	return nil
-
-}
-
-// QueryBucketItemCount uses a request plus query to get the number of items in a bucket, as the REST API can be slow to update its value.
-// Requires a primary index on the bucket.
-func QueryBucketItemCount(n1qlStore N1QLStore) (itemCount int, err error) {
-	statement := fmt.Sprintf("SELECT COUNT(1) AS count FROM %s", KeyspaceQueryToken)
-	r, err := n1qlStore.Query(statement, nil, RequestPlus, true)
-	if err != nil {
-		return -1, err
-	}
-	var val struct {
-		Count int `json:"count"`
-	}
-	err = r.One(&val)
-	if err != nil {
-		return -1, err
-	}
-	return val.Count, nil
-}
-
-func isMinimumVersion(major, minor, requiredMajor, requiredMinor uint64) bool {
-	if major < requiredMajor {
-		return false
-	}
-
-	if major == requiredMajor && minor < requiredMinor {
-		return false
-	}
-
-	return true
-}
-
-func normalizeIntToUint(value interface{}) (uint, error) {
-	switch typeValue := value.(type) {
-	case int:
-		return uint(typeValue), nil
-	case uint64:
-		return uint(typeValue), nil
-	case string:
-		i, err := strconv.Atoi(typeValue)
-		return uint(i), err
-	default:
-		return uint(0), fmt.Errorf("Unable to convert %v (%T) -> uint.", value, value)
-	}
-}
-
-func asBool(value interface{}) bool {
-
-	switch typeValue := value.(type) {
-	case string:
-		parsedVal, err := strconv.ParseBool(typeValue)
-		if err != nil {
-			WarnfCtx(context.Background(), "asBool called with unknown value: %v.  defaulting to false", typeValue)
-			return false
-		}
-		return parsedVal
-	case bool:
-		return typeValue
-	default:
-		WarnfCtx(context.Background(), "asBool called with unknown type: %T.  defaulting to false", typeValue)
-		return false
-	}
-
-}
-
-// AsLeakyBucket tries to return the given bucket as a LeakyBucket.
-func AsLeakyBucket(bucket Bucket) (*LeakyBucket, bool) {
-
-	var underlyingBucket Bucket
-	switch typedBucket := bucket.(type) {
-	case *LeakyBucket:
-		return typedBucket, true
-	case *TestBucket:
-		underlyingBucket = typedBucket.Bucket
-	default:
-		// bail out for unrecognised/unsupported buckets
-		return nil, false
-	}
-
-	return AsLeakyBucket(underlyingBucket)
-}
-
-func GoCBBucketMgmtEndpoints(bucket CouchbaseBucketStore) (url []string, err error) {
-	return bucket.MgmtEps()
-}
-
-// Get one of the management endpoints.  It will be a string such as http://couchbase
-func GoCBBucketMgmtEndpoint(bucket CouchbaseBucketStore) (url string, err error) {
-	mgmtEps, err := bucket.MgmtEps()
-	if err != nil {
-		return "", err
-	}
-	bucketEp := mgmtEps[rand.Intn(len(mgmtEps))]
-	return bucketEp, nil
 }
